requirements.txt
Flask==1.1.2
boto3==1.16.5
opencv-python==4.4.0.44
numpy==1.18.5
Pillow==8.0.1